fix compatibility with framework
